#95 fix++

+++ dwarf_debugger/ui/dialogs/dialog_scripts.py
@@ -92,7 +92,7 @@ class ScriptsDialog(QDialog):
         self._new_pixmap = QPixmap(max(_section_width, 40), 20)
         self._new_pixmap.fill(Qt.transparent)
         painter = QPainter(self._new_pixmap)
-        rect = QRect((_section_width * 0.5) - 5, 0, 20, 20)
+        rect = QRect(int(_section_width * 0.5) - 5, 0, 20, 20)
         painter.setBrush(QColor('#666'))
         painter.setPen(QColor('#666'))
         painter.drawEllipse(rect)
